requirements.txt
kagglehub
tensorflow
opencv-python
numpy
pyttsx3
flask
pillow
duckduckgo-search
requests
Pillow